+++ src/index.ts
@@ -5,4 +5,4 @@ export * from './common/interfaces/indexes/index';
 
 export * from './components/interfaces/indexes/index';
 export * from './logic/indexes/index';
-export * from './pnpjs';
+export * from './pnpjs/index';