+++ CHANGELOG.md
@@ -12,3 +12,4 @@ npm install @mikezimm/fps-library-v2@1.0.0
 # Changelog
 
 ## 1.0.0 - 2022-Dec-09
+- completely rebuilt userServices in src\pnpjs\Users